Help messages retriever: use a list to build all sorts of help messages, along iwth invoking object class method resolution order.
As a first step in redesigning Control Usage Assistant, a help messages list will be used to gather help messages from everywhere, including object's method resolution order 9MRO). By using MRO, it allows objects to provide help messages in the future, beginning with more specific help messages for the currently active overlay class to a more generic ones for a family of objects (for example, an edit field in an app module, a message for interacting with edit fields, in that order). The previous work on rolles will not be forgotten - these must be consulted in the end.